Function that sums all its parameters

+++ 0x10-variadic_functions/0-sum_them_all.c
@@ -10,7 +10,7 @@ int sum_them_all(const unsigned int n, ...)
 {
 	va_list args;
 	unsigned int i;
-	long int sum;
+	int sum;
 
 	if (!n)
 		return (0);
